Implement a Proj extension
This implements a package extension for Proj and avoids us having to depend explicitly on it.  With this, GeometryOps has no explicit dependencies on C libraries!  We might also extend the `reproject` interface to allow e.g. FastGeoProjections and other projection interfaces.

The issue with this is that the GeodesicSegments struct holds a Proj.geod_geodesic object.  That field is currently untyped, we may choose to type it and only accept GeodesicSegments{Proj.geod_geodesic} in principle.