Fix touches for multi geoms
Only one of the geoms in g2 MUST touch the geom in g1, all others MUST be disjoint or touching.

This could potentially be accelerated by building an STRtree if the query width is long enough, or we may also need some tree interface / preparation here.  E.g. Canada has a lot of islands that a tree approach could get rid of.