Basic implementation of spatial tree interface
This is only really workable for (assumed static) trees...but very powerful abstraction.  Assumes a certain structure to all spatial trees which may not be desirable.  We may also want to combine trees with geometry representations in preparations - which this allows us to do but requires some acrobatics.

Notably this interface does NOT help us get the actual geometries, trees are assumed to store indices to geometries.  Maybe that's a bad idea?  But considering featurecollections etc it seems to be the most prudent thing.