Create a TracingError type that stores all context
Better than printing out since we can (A) check on it and (B) access the polygons as well as the a_list and b_list in Julia.  I have a few plot recipes for them but not sure if they are worth putting in a Makie extension...they would also be inaccessible there.  Maybe in GeoMakie?  Then we can make full use of customizations etc.