Implement robust cross product
Ported the implementation from s2.  This is the core component in spherical intersection_point which we will need for polygon intersection.  Thankfully with a few tweaks, FosterHormannClipping should permit spherical points as well and I don't think we need to make any changes after that.  Predicates should also work in 3d, maybe with an implicit (0, 0, 0) point to round out the plane.